requirements.txt
# Core
pyyaml>=6.0
pydantic>=2.0.0

# Orchestrator & Agents
openai>=1.0.0
anthropic>=0.18.0
httpx>=0.25.0
aiohttp>=3.9.0

# Integrations
python-telegram-bot>=21.0

# Dashboard
fastapi>=0.109.0
uvicorn[standard]>=0.27.0
python-multipart>=0.0.6
jinja2>=3.1.0
websockets>=12.0
passlib[bcrypt]>=1.7.4

# Database & Memory
chromadb>=0.4.0

# Tools
playwright>=1.40.0

# Scheduling
apscheduler>=3.10.0
